Split the resampling module
Future changes require adding more code to this module, which is already
quite big. This commit makes it a package and split the module into a
few sub-modules to make it more manageable.
