+++ src/frequenz/sdk/timeseries/_resampling/__init__.py
@@ -1,0 +1,4 @@
+# License: MIT
+# Copyright © 2024 Frequenz Energy-as-a-Service GmbH
+
+"""Timeseries resampling package."""
